requirements.txt
Flask==2.0.1
torch==2.2.1
numpy==1.20.3
einops==0.3.0
gtts==2.2.3
playsound==1.2.2
python-dotenv==0.17.1
openai==0.10.2
